When transitions update a state, they can also return relevant information about the transition. This lets the underlying model (such as the LSTMModel) handle different versions of the same transition differently. For example, Shift(word) can have a different embedding from a generic Shift in the transition embedding.
Rough draft of common words shift transitions.  Need to test it some more

Attempt to label the close transitions as well (turned off, wasn't working)

Includes a sanity check for the added transitions

Average the specialized vector and the base vector when a labeled shift or close is used

The transition_embedding labels on Shift and Close are a command line option.  Will help A/B testing if nothing else...

Try tags instead of common words?